refactor: extract command processing to dedicated methods
#15
