+++ cli.js
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 
 const program = require('commander')
-const { generateComponentFiles, generateFilesIfNotExistAlready} = require('./src/gulpfile')
 const fs = require('fs')
 const path = require('path')
 const promptly = require('promptly')
 const { readFileSync } = require('fs')
 const { copySync } = require('fs-extra')
-const { getDirectories, validateKebabCaseName } = require('./src/utilities')
+const { getDirectories } = require('./src/utilities')
+const {processPromptCommand, processUpgradeCommand, processCreateComponentCommand} = require('./src/commands')
 
 const CONFIG_DIRECTORY = '.create-frontend-component'
 const CONFIG_FILE_NAME = 'config.json'
@@ -22,7 +22,6 @@ const configDefaults = {
 
 
 /**
- * @param {string} filePath 
  * @return {object}
  */
 function loadConfig() {
@@ -74,7 +73,7 @@ program
   .arguments('<component-name>')
   .option( '-t, --type <type>', 'Component type, default: atoms')
   .option( '-f, --flavour <flavour>', 'Component flavour')
-  .action( function(componentName, env) {
+  .action( async function(componentName, env) {
     if (componentName.toLowerCase() === 'init') {
       const availablePresets = getDirectories(PRESET_PATH)
       promptly.choose('Choose a preset (' + availablePresets.join(', ') + '): ', availablePresets).then(
@@ -89,84 +88,11 @@ program
     const availableFlavours = getDirectories(fullTemplatePath)
 
     if (componentName.toLowerCase() === 'prompt') {
-      const context = {}
-      promptly.prompt('Component Name (kebab-case): ', { validator: validateKebabCaseName }).then(
-        (componentName) => {
-          context.componentName = componentName
-          return promptly.choose('Choose a type (' + allowedComponentTypes.join(', ') + '): ', allowedComponentTypes)
-        }
-      ).then(
-        (componentType) => {
-          context.componentType = componentType
-
-          if (availableFlavours.length === 0) {
-            console.warn('Could not detect any component flavour, falling back to "default"')
-            return Promise.resolve('default')
-          }
-
-          if (availableFlavours.length === 1) {
-            return Promise.resolve(availableFlavours[0])
-          }
-
-          return promptly.choose('Choose a flavour (' + availableFlavours.join(', ') + '): ', availableFlavours)
-        }
-      ).then(
-        (flavour) => {
-          generateComponentFiles(fullTemplatePath, componentPath, context.componentName, context.componentType, flavour, availableFlavours)
-          return
-        }
-      )
-      return
-    }
-
-    if (componentName.toLowerCase() === 'upgrade') {
-      const context = {}
-      if (availableFlavours.length <= 1) {
-        console.warn('Could not detect more than 1 flavour, upgrade is not possible')
-        return
-      }
-
-      promptly.prompt('Component Name (kebab-case): ', { validator: validateKebabCaseName })
-        .then(
-          (componentName) => {
-            context.componentName = componentName
-            return promptly.choose('Which type is your component? (' + allowedComponentTypes.join(', ') + '): ', allowedComponentTypes)
-          }
-        )
-        .then(
-          (componentType) => {
-            context.componentType = componentType
-
-            return promptly.choose('Choose a flavour to upgrade (' + availableFlavours.join(', ') + '): ', availableFlavours)
-          }
-        ).then(
-          (flavour) => {
-            generateFilesIfNotExistAlready(fullTemplatePath, componentPath, context.componentName, context.componentType, flavour, availableFlavours)
-            return
-          }
-        )
-      return
-    }
-
-
-    if (env.type && allowedComponentTypes.length == 0) {
-      throw new Error('component types are not configured in this project but found parameter "type"')
-    }
-
-    let componentType = env.type
-    if (allowedComponentTypes.length === 0) {
-      componentType = null
-    } else if (!componentType) {
-      // use first type as default
-      componentType = allowedComponentTypes[0]
+      await processPromptCommand(allowedComponentTypes, availableFlavours, fullTemplatePath, componentPath)
+    } else if (componentName.toLowerCase() === 'upgrade') {
+      await processUpgradeCommand(availableFlavours, allowedComponentTypes, fullTemplatePath, componentPath)
     } else {
-      componentType = componentType.toLowerCase()
+      processCreateComponentCommand(env, allowedComponentTypes, fullTemplatePath, componentPath, componentName, availableFlavours)
     }
-
-    if (componentType && !allowedComponentTypes.includes(componentType)) {
-      throw new Error(`component type '${componentType}' is not allowed, choose one of: ${allowedComponentTypes}`)
-    }
-
-    generateComponentFiles(fullTemplatePath, componentPath, componentName, componentType, env.flavour, availableFlavours)
   })
   .parse(process.argv)
